+++ packages/components/src/components/Dialog/styles/dialogStyles.ts
